+++ src/index.ts
@@ -1,7 +1,9 @@
 import search from "./search";
+import random from "./random";
 import { resetCache } from "./spellbook-api";
 
 export = {
   search,
+  random,
   resetCache,
 };